GCA-50 Add Hibernate mapping for Entities

+++ src/main/java/com/gcs/app/facade/dto/TrainingResponseDto.java
@@ -13,5 +13,5 @@ public class TrainingResponseDto {
     private String name;
     private TrainingType type;
     private LocalDate date;
-    private Double duration;
+    private Long duration;
 }
